+++ java/hibernate/examples/pet-clinic-app/src/test/java/org/springframework/samples/petclinic/owner/PetControllerTests.java
@@ -29,11 +29,11 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.condition.DisabledInNativeImage;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.boot.webmvc.test.autoconfigure.WebMvcTest;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.FilterType;
 import org.springframework.test.context.aot.DisabledInAotMode;
+import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 
 /**
@@ -55,7 +55,7 @@ class PetControllerTests {
 
   @Autowired private MockMvc mockMvc;
 
-  @MockBean private OwnerRepository owners;
+  @MockitoBean private OwnerRepository owners;
 
   @BeforeEach
   void setup() {