+++ java/hibernate/examples/pet-clinic-app/src/test/java/org/springframework/samples/petclinic/system/CrashControllerIntegrationTests.java
@@ -25,11 +25,12 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceTransactionManagerAutoConfiguration;
-import org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration;
+import org.springframework.boot.hibernate.autoconfigure.HibernateJpaAutoConfiguration;
+import org.springframework.boot.jdbc.autoconfigure.DataSourceAutoConfiguration;
+import org.springframework.boot.jdbc.autoconfigure.DataSourceTransactionManagerAutoConfiguration;
+import org.springframework.boot.resttestclient.TestRestTemplate;
+import org.springframework.boot.resttestclient.autoconfigure.AutoConfigureTestRestTemplate;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
@@ -48,9 +49,10 @@ import org.springframework.http.ResponseEntity;
 @SpringBootTest(
     webEnvironment = RANDOM_PORT,
     properties = {
-      "server.error.include-message=ALWAYS",
-      "management.endpoints.enabled-by-default=false"
+      "spring.web.error.include-message=ALWAYS",
+      "management.endpoints.access.default=none"
     })
+@AutoConfigureTestRestTemplate
 class CrashControllerIntegrationTests {
 
   @SpringBootApplication(